requirements.txt
setuptools
werkzeug
flask
flask-httpauth
flask-socketio
eventlet
sqlalchemy
requests
decorator
dnspython
ipaddress
simpleeval
pyyaml
argon2
